# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JAC Hub API - Panel de gestión para estudio indie de videojuegos
servers:
  - url: /api
    description: Ruta base de la API
...
paths:
...
  /proyectos:
    get:
      operationId: listProyectos
      tags: [proyectos]
      summary: Listar proyectos
      responses:
        "200":
          description: Lista de proyectos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proyecto"
    post:
      operationId: createProyecto
      tags: [proyectos]
      summary: Crear proyecto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProyectoInput"
      responses:
        "201":
          description: Proyecto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proyecto"
  /proyectos/{id}:
    get:
      operationId: getProyecto
      tags: [proyectos]
      summary: Obtener proyecto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proyecto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proyecto"
        "404":
          description: Proyecto no encontrado
    patch:
      operationId: updateProyecto
      tags: [proyectos]
      summary: Actualizar proyecto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProyectoUpdate"
      responses:
        "200":
          description: Proyecto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proyecto"
    delete:
      operationId: deleteProyecto
      tags: [proyectos]
      summary: Eliminar proyecto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Proyecto eliminado
  /tareas:
    get:
      operationId: listTareas
      tags: [tareas]
      summary: Listar tareas
      parameters:
        - name: proyectoId
          in: query
          schema:
            type: integer
        - name: columna
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tarea"
    post:
      operationId: createTarea
      tags: [tareas]
      summary: Crear tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TareaInput"
      responses:
        "201":
          description: Tarea creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarea"
  /tareas/{id}:
    patch:
      operationId: updateTarea
      tags: [tareas]
      summary: Actualizar tarea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TareaUpdate"
      responses:
        "200":
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarea"
    delete:
      operationId: deleteTarea
      tags: [tareas]
      summary: Eliminar tarea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Tarea eliminada
...
components:
  schemas:
...
    Proyecto:
      type: object
      required: [id, nombre, estado, progreso, descripcion, fechaCreacion, fechaActualizacion]
      properties:
        id:
          type: integer
        nombre:
          type: string
        estado:
          type: string
        progreso:
          type: integer
        descripcion:
          type: string
        fechaCreacion:
          type: string
        fechaActualizacion:
          type: string
        miembros:
          type: array
          items:
            type: string
        tecnologias:
          type: array
          items:
            type: string
    ProyectoInput:
      type: object
      required: [nombre, descripcion]
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        estado:
          type: string
        tecnologias:
          type: array
          items:
            type: string
    ProyectoUpdate:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        estado:
          type: string
        progreso:
          type: integer
        tecnologias:
          type: array
          items:
            type: string
    Tarea:
      type: object
      required: [id, titulo, columna, prioridad, fechaCreacion]
      properties:
        id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: ["string", "null"]
        columna:
          type: string
        prioridad:
          type: string
        proyectoId:
          type: ["integer", "null"]
        asignadoA:
          type: ["string", "null"]
        creadoPor:
          type: ["string", "null"]
        etiquetas:
          type: array
          items:
            type: string
        fechaCreacion:
          type: string
        fechaLimite:
          type: ["string", "null"]
    TareaInput:
      type: object
      required: [titulo, columna]
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        columna:
          type: string
        prioridad:
          type: string
        proyectoId:
          type: integer
        asignadoA:
          type: string
        creadoPor:
          type: string
        etiquetas:
          type: array
          items:
            type: string
        fechaLimite:
          type: string
    TareaUpdate:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        columna:
          type: string
        prioridad:
          type: string
        asignadoA:
          type: string
        etiquetas:
          type: array
          items:
            type: string
        fechaLimite:
          type: string
...